Add version to AssociatedFile metadata
AdServices, which uses BertNLClassifier for TFLite model inference,
needs to maintain separate versioning for the model itself and its
taxonomy/labels.

This CL adds a version field to the AssociatedFile
metadata schema, which maintains metadata for files such as labels and
vocabulary.

It also adds a new method to access a given AssociatedFile's metadata
by name, and standardizes the signature of GetModelVersion to match.

Finally, it saves a copy of a given NLClassifier's labels metadata
upon loading the labels file and adds methods to get the labels version.

Bug: 232807230, Bug: 232807776
Test: atest BertNLClassifierTest
Change-Id: If613972c5ea9e30a14fbf478f3eca3f299dd1bd1