Return StatusOr from Classify() methods
This CL updates the Classify() methods in NLClassifier and
BertNLClassifier to return StatusOr wrappers for their contents,
allowing errors to be propagated across the JNI boundary.

To enable this, it fixes a bug when checking for dynamic vs static input
tensors in bert_nl_classifier.cc and throws exceptions from
RunClassifier().

Bug: 242926638, Bug: 242926783
Test: atest tflite_support_classifier_tests
Test: atest TfliteSupportClassifierTests
Change-Id: Ib93385061ace97cf96a168e3d76d2d9162c8055f